Create user-switching-1.php
I use this plugin at front end for the application level administrators. It is more convenient for them when switch back to own account return not to the current page, but to the own admin page, which is not available for users. So I added filter to the action_wp_footer() function, which allow to replace 'redirect_to' parameter.